pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "duplifinder"
version = "2.12.0-alpha.1"
description = "Find duplicate Python definitions across projects"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "you@example.com" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]
keywords = ["duplicates", "code-analysis", "python", "refactoring"]

dependencies = [
    "pyyaml>=6.0",
    "rich>=12.0; sys_platform != 'win32'",
    "tqdm>=4.0; sys_platform != 'win32'",
    "pydantic>=2.0"  # For config validation
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
    "pytest-cov>=4.0",
    "pytest-mock>=3.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/duplifinder"
Repository = "https://github.com/yourusername/duplifinder"
Issues = "https://github.com/yourusername/duplifinder/issues"

[project.scripts]
duplifinder = "duplifinder.main:main"
